REFACTOR(competition): DEDUPLICATE ENROLLMENT EXCEPTIONS AND TEST HELPERS
Move EnrollmentNotFoundError and NotCreatorError into the shared
application/exceptions.py module, removing the duplicate class
definitions previously redefined in each enrollment use case
(set/remove custom handicap, withdraw, cancel, handle, direct enroll).
Also extract the duplicated competition/enrollment test fixtures
shared by the set/remove custom handicap use case tests into a
common helpers module.
