+++ src/modules/competition/application/exceptions.py
@@ -107,3 +107,15 @@ class HandicapEditNotAllowedError(Exception):
     """El hándicap personalizado no puede modificarse en el estado actual de la competición."""
 
     pass
+
+
+class EnrollmentNotFoundError(Exception):
+    """La inscripción no existe."""
+
+    pass
+
+
+class NotCreatorError(Exception):
+    """El usuario no es el creador de la competición."""
+
+    pass
