+++ src/modules/competition/application/use_cases/cancel_enrollment_use_case.py
@@ -8,17 +8,12 @@ from src.modules.competition.application.dto.enrollment_dto import (
     CancelEnrollmentRequestDTO,
     CancelEnrollmentResponseDTO,
 )
+from src.modules.competition.application.exceptions import EnrollmentNotFoundError
 from src.modules.competition.domain.repositories.competition_unit_of_work_interface import (
     CompetitionUnitOfWorkInterface,
 )
 from src.modules.competition.domain.value_objects.enrollment_id import EnrollmentId
 from src.modules.user.domain.value_objects.user_id import UserId
-
-
-class EnrollmentNotFoundError(Exception):
-    """Excepción lanzada cuando la inscripción no existe."""
-
-    pass
 
 
 class NotOwnerError(Exception):
