+++ src/modules/competition/application/use_cases/handle_enrollment_use_case.py
@@ -8,25 +8,17 @@ from src.modules.competition.application.dto.enrollment_dto import (
     HandleEnrollmentRequestDTO,
     HandleEnrollmentResponseDTO,
 )
-from src.modules.competition.application.exceptions import CompetitionNotFoundError
+from src.modules.competition.application.exceptions import (
+    CompetitionNotFoundError,
+    EnrollmentNotFoundError,
+    NotCreatorError,
+)
 from src.modules.competition.domain.repositories.competition_unit_of_work_interface import (
     CompetitionUnitOfWorkInterface,
 )
 from src.modules.competition.domain.services.competition_policy import CompetitionPolicy
 from src.modules.competition.domain.value_objects.enrollment_id import EnrollmentId
 from src.modules.user.domain.value_objects.user_id import UserId
-
-
-class EnrollmentNotFoundError(Exception):
-    """Excepción lanzada cuando la inscripción no existe."""
-
-    pass
-
-
-class NotCreatorError(Exception):
-    """Excepción lanzada cuando el usuario no es el creador de la competición."""
-
-    pass
 
 
 class InvalidActionError(Exception):
