+++ tests/unit/modules/competition/application/use_cases/test_remove_custom_handicap_use_case.py
@@ -1,20 +1,13 @@
 """Tests para RemoveCustomHandicapUseCase."""
 
-from datetime import date
 from decimal import Decimal
 from uuid import uuid4
 
 import pytest
 
-from src.modules.competition.application.dto.competition_dto import (
-    CreateCompetitionRequestDTO,
-)
 from src.modules.competition.application.exceptions import (
     CompetitionNotFoundError,
     HandicapEditNotAllowedError,
-)
-from src.modules.competition.application.use_cases.create_competition_use_case import (
-    CreateCompetitionUseCase,
 )
 from src.modules.competition.application.use_cases.remove_custom_handicap_use_case import (
     EnrollmentNotFoundError,
@@ -28,6 +21,11 @@ from src.modules.competition.infrastructure.persistence.in_memory.in_memory_unit
     InMemoryUnitOfWork,
 )
 from src.modules.user.domain.value_objects.user_id import UserId
+from tests.unit.modules.competition.application.use_cases.helpers import (
+    create_approved_enrollment,
+    create_competition,
+    set_competition_status,
+)
 
 pytestmark = pytest.mark.asyncio
 
@@ -51,53 +49,6 @@ class TestRemoveCustomHandicapUseCase:
     def admin_user_id(self) -> UserId:
         return UserId(uuid4())
 
-    async def _create_competition(self, uow: InMemoryUnitOfWork, creator_id: UserId):
-        """Helper: crea una competición en DRAFT."""
-        create_uc = CreateCompetitionUseCase(uow)
-        request = CreateCompetitionRequestDTO(
-            name="Test Cup",
-            start_date=date(2026, 6, 1),
-            end_date=date(2026, 6, 3),
-            main_country="ES",
-            play_mode="SCRATCH",
-            max_players=24,
-        )
-        return await create_uc.execute(request, creator_id)
-
-    async def _set_competition_status(
-        self, uow: InMemoryUnitOfWork, competition_id, status: str
-    ):
-        """Helper: mueve una competición a través de sus transiciones hasta `status`."""
-        async with uow:
-            competition = await uow.competitions.find_by_id(CompetitionId(competition_id))
-            if status in ("ACTIVE", "CLOSED", "IN_PROGRESS", "COMPLETED", "CANCELLED"):
-                competition.activate()
-            if status in ("CLOSED", "IN_PROGRESS", "COMPLETED"):
-                competition.close_enrollments()
-            if status in ("IN_PROGRESS", "COMPLETED"):
-                competition.start()
-            if status == "COMPLETED":
-                competition.complete()
-            if status == "CANCELLED":
-                competition.cancel()
-            await uow.competitions.update(competition)
-            await uow.commit()
-
-    async def _create_enrollment_with_custom_handicap(
-        self, uow: InMemoryUnitOfWork, competition_id, user_id: UserId
-    ) -> Enrollment:
-        """Helper: crea un enrollment aprobado con hándicap personalizado ya establecido."""
-        enrollment = Enrollment.direct_enroll(
-            id=EnrollmentId.generate(),
-            competition_id=CompetitionId(competition_id),
-            user_id=user_id,
-            custom_handicap=Decimal("20.0"),
-        )
-        async with uow:
-            await uow.enrollments.add(enrollment)
-            await uow.commit()
-        return enrollment
-
     async def test_should_remove_custom_handicap_successfully(
         self, uow: InMemoryUnitOfWork, creator_id: UserId, player_id: UserId
     ):
@@ -106,11 +57,11 @@ class TestRemoveCustomHandicapUseCase:
         When: El creador lo elimina
         Then: El enrollment vuelve a tener custom_handicap=None
         """
-        created = await self._create_competition(uow, creator_id)
-        enrollment = await self._create_enrollment_with_custom_handicap(
-            uow, created.id, player_id
+        created = await create_competition(uow, creator_id)
+        enrollment = await create_approved_enrollment(
+            uow, created.id, player_id, custom_handicap=Decimal("20.0")
         )
-        await self._set_competition_status(uow, created.id, "ACTIVE")
+        await set_competition_status(uow, created.id, "ACTIVE")
 
         use_case = RemoveCustomHandicapUseCase(uow)
         response = await use_case.execute(str(enrollment.id.value), creator_id)
@@ -129,11 +80,11 @@ class TestRemoveCustomHandicapUseCase:
         When: Un admin (no creador) lo elimina con is_admin=True
         Then: Se aplica sin lanzar NotCreatorError
         """
-        created = await self._create_competition(uow, creator_id)
-        enrollment = await self._create_enrollment_with_custom_handicap(
-            uow, created.id, player_id
+        created = await create_competition(uow, creator_id)
+        enrollment = await create_approved_enrollment(
+            uow, created.id, player_id, custom_handicap=Decimal("20.0")
         )
-        await self._set_competition_status(uow, created.id, "ACTIVE")
+        await set_competition_status(uow, created.id, "ACTIVE")
 
         use_case = RemoveCustomHandicapUseCase(uow)
         response = await use_case.execute(
@@ -150,11 +101,11 @@ class TestRemoveCustomHandicapUseCase:
         When: Un usuario que no es el creador ni admin intenta eliminarlo
         Then: Se lanza NotCreatorError
         """
-        created = await self._create_competition(uow, creator_id)
-        enrollment = await self._create_enrollment_with_custom_handicap(
-            uow, created.id, player_id
+        created = await create_competition(uow, creator_id)
+        enrollment = await create_approved_enrollment(
+            uow, created.id, player_id, custom_handicap=Decimal("20.0")
         )
-        await self._set_competition_status(uow, created.id, "ACTIVE")
+        await set_competition_status(uow, created.id, "ACTIVE")
 
         use_case = RemoveCustomHandicapUseCase(uow)
 
@@ -206,8 +157,8 @@ class TestRemoveCustomHandicapUseCase:
         When: El creador intenta eliminar el hándicap
         Then: Se lanza EnrollmentStateError
         """
-        created = await self._create_competition(uow, creator_id)
-        await self._set_competition_status(uow, created.id, "ACTIVE")
+        created = await create_competition(uow, creator_id)
+        await set_competition_status(uow, created.id, "ACTIVE")
         enrollment = Enrollment.request(
             id=EnrollmentId.generate(),
             competition_id=CompetitionId(created.id),
@@ -232,11 +183,11 @@ class TestRemoveCustomHandicapUseCase:
         When: El creador intenta eliminar el hándicap
         Then: Se lanza HandicapEditNotAllowedError
         """
-        created = await self._create_competition(uow, creator_id)
-        enrollment = await self._create_enrollment_with_custom_handicap(
-            uow, created.id, player_id
+        created = await create_competition(uow, creator_id)
+        enrollment = await create_approved_enrollment(
+            uow, created.id, player_id, custom_handicap=Decimal("20.0")
         )
-        await self._set_competition_status(uow, created.id, status)
+        await set_competition_status(uow, created.id, status)
 
         use_case = RemoveCustomHandicapUseCase(uow)
 
@@ -252,11 +203,11 @@ class TestRemoveCustomHandicapUseCase:
         When: La competición está en DRAFT, ACTIVE o CLOSED
         Then: Se puede eliminar el hándicap personalizado sin errores
         """
-        created = await self._create_competition(uow, creator_id)
-        enrollment = await self._create_enrollment_with_custom_handicap(
-            uow, created.id, player_id
+        created = await create_competition(uow, creator_id)
+        enrollment = await create_approved_enrollment(
+            uow, created.id, player_id, custom_handicap=Decimal("20.0")
         )
-        await self._set_competition_status(uow, created.id, status)
+        await set_competition_status(uow, created.id, status)
 
         use_case = RemoveCustomHandicapUseCase(uow)
         response = await use_case.execute(str(enrollment.id.value), creator_id)
